% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fun.R
\name{norm1}
\alias{norm1}
\title{auxilary code
please load this function into your R-console by copy and paste
RECA will call this function during the modeling process}
\usage{
norm1(x)
}
\description{
auxilary code
please load this function into your R-console by copy and paste
RECA will call this function during the modeling process
}
