% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_matrix.R
\name{tcalc}
\alias{tcalc}
\title{T-matrix calculation see Weltje, 1997 Appendix C, k=matrix M, c=weighting Factor}
\usage{
tcalc(k, c)
}
\description{
T-matrix calculation see Weltje, 1997 Appendix C, k=matrix M, c=weighting Factor
}
